--- src/aqua.jl
@@ -1,1 +1,0 @@
-# suggestion from Aqua.test_all(BaseFloats) to resolve ambiguities
